Refactoring and dumping more documentation to the Readme

+++ geo/const.go
@@ -13,7 +13,7 @@ const (
 	//
 	// Map related
 	//
-	TileSizeInPixels  int     = 256               // Tile size 256x256
+	TileSizeInPixels  float64 = 256.0             // Tile size is 256x256
 	TileSizeInRadians float64 = 128 / math.Pi     // A value that represents a map that extends from 0 to 256 / (2 * PI)
 	LonEdge           float64 = 180               // From −180° to 180° Longitude
 	LatEdge           float64 = 85.05112877980659 // Also know as yMax -85.05° North and 85.05° South
